requirements.txt
# LLM APIs
openai==1.54.0
anthropic==0.39.0

# LangChain (for orchestration)
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.3.0
langgraph==0.2.45

# FastAPI
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Streamlit
streamlit==1.40.2

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Redis
redis==5.2.0

# S3 / MinIO
boto3==1.35.70
minio==7.2.10

# Vector DB
chromadb==0.5.23

# Embeddings
sentence-transformers==3.3.1

# Monitoring
langfuse==2.56.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
pandas==2.2.3

# Email
python-smtplib==0.1.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0